+++ src/CashFlow.Application/useCases/Expenses/Delete/IDeleteExpenseUseCase.cs
@@ -1,5 +1,5 @@
 ﻿namespace CashFlow.Application.useCases.Expenses.Delete;
 public interface IDeleteExpenseUseCase
 {
-    Task Execute(int id);
+    Task Execute(long id);
 }
