+++ src/CashFlow.Application/useCases/Expenses/Update/IUpdateExpenseUseCase.cs
@@ -3,5 +3,5 @@
 namespace CashFlow.Application.useCases.Expenses.Update;
 public interface IUpdateExpenseUseCase
 {
-    Task Execute(int id, RequestExpenseJson request );
+    Task Execute(long id, RequestExpenseJson request );
 }
