+++ src/CashFlow.Domain/Repositories/Expenses/IExpensesReadOnlyRepository.cs
@@ -4,6 +4,6 @@ namespace CashFlow.Domain.Repositories.Expenses;
 public interface IExpensesReadOnlyRepository
 {
     public Task<List<Expense>> GetAll(User user);
-    public Task<Expense?> GetById(int id, User user);
+    public Task<Expense?> GetById(long id, User user);
     public Task<List<Expense>> FilterByMonth(User user,DateOnly month);
 }
