+++ src/CashFlow.Domain/Repositories/Expenses/IExpensesUpdateOnlyRepository.cs
@@ -3,6 +3,6 @@
 namespace CashFlow.Domain.Repositories.Expenses;
 public interface IExpensesUpdateOnlyRepository
 {
-    Task<Expense?> GetById(User user,int id);
+    Task<Expense?> GetById(User user,long id);
     void Update(Expense expense);
 }
